+++ pocketutils/misc/resources.py
@@ -1,60 +1,95 @@
 import logging
-import os
 import typing
-from dataclasses import dataclass
 from datetime import datetime, timedelta
 from pathlib import Path
-from typing import MutableMapping, Optional, Any, Union
+from typing import MutableMapping, Optional, Union
 
 import orjson
+
+from pocketutils.core import PathLike
 from pocketutils.core.chars import Chars
 from pocketutils.core.dot_dict import NestedDotDict
-from pocketutils.core.exceptions import FileDoesNotExistError, MissingResourceError, PathExistsError
+from pocketutils.core.exceptions import (
+    FileDoesNotExistError,
+    MissingResourceError,
+    PathExistsError,
+    DirDoesNotExistError,
+)
 from pocketutils.tools.common_tools import CommonTools
 from pocketutils.tools.unit_tools import UnitTools
+from pocketutils.tools.filesys_tools import FilesysTools
 
 
-logger = logging.getLogger("pocketutils")
+_logger = logging.getLogger("pocketutils")
 
 
-@dataclass(frozen=True, repr=True)
-class MandosResources:
-    resource_dir: Path
-    logger: Any = logger
+class Resources:
+    def __init__(self, path: PathLike, *, logger=_logger):
+        self._dir = Path(path)
+        self._logger = logger
 
-    def contains(self, *nodes: Union[Path, str], suffix: Optional[str] = None) -> bool:
+    def contains(self, *nodes: PathLike, suffix: Optional[str] = None) -> bool:
         """Returns whether a resource file (or dir) exists."""
         return self.path(*nodes, suffix=suffix).exists()
 
-    def path(
-        self, *nodes: Union[Path, str], suffix: Optional[str] = None, exists: bool = False
-    ) -> Path:
-        """Gets a path of a test resource file under ``resources/``."""
-        path = Path(self.resource_dir, "resources", *nodes)
+    def path(self, *nodes: PathLike, suffix: Optional[str] = None, exists: bool = False) -> Path:
+        """
+        Gets a path of a test resource file under ``resources/``.
+
+        Raises:
+            MissingResourceError: If the path is not found
+        """
+        path = Path(self._dir, "resources", *nodes)
         path = path.with_suffix(path.suffix if suffix is None else suffix)
         if exists and not path.exists():
             raise MissingResourceError(f"Resource {path} missing")
         return path
 
-    def file(self, *nodes: Union[Path, str], suffix: Optional[str] = None) -> Path:
-        """Gets a path of a test resource file under ``resources/``."""
+    def file(self, *nodes: PathLike, suffix: Optional[str] = None) -> Path:
+        """
+        Gets a path of a test resource file under ``resources/``.
+
+        Raises:
+            MissingResourceError: If the path is not found
+            PathExistsError: If the path is not a file or symlink to a file,
+                             or is not readable
+        """
         path = self.path(*nodes, suffix=suffix)
-        if not path.is_file():
+        info = FilesysTools.get_info(path)
+        if not info.is_file:
             raise PathExistsError(f"Resource {path} is not a file!")
-        if not os.access(path, os.R_OK):
+        if not info.is_readable_file:
             raise FileDoesNotExistError(f"Resource {path} is not readable")
         return path
 
-    def dir(self, *nodes: Union[Path, str]) -> Path:
-        """Gets a path of a test resource file under ``resources/``."""
+    def dir(self, *nodes: PathLike) -> Path:
+        """
+        Gets a path of a test resource file under ``resources/``.
+
+        Raises:
+            MissingResourceError: If the path is not found and ``not missing_ok``
+            PathExistsError: If the path is not a dir, symlink to a dir, or mount,
+                             or lacks 'R' or 'X' permissions
+        """
         path = self.path(*nodes)
-        if not path.is_dir() and not path.is_mount():
-            raise PathExistsError(f"Resource {path} is not a directory!")
+        info = FilesysTools.get_info(path)
+        if not info.exists:
+            raise DirDoesNotExistError(f"Resource {path} does not exist")
+        if not info.is_dir:
+            raise PathExistsError(f"Resource {path} is not a directory")
+        if info.is_readable_dir:
+            raise FileDoesNotExistError(f"Resource {path} is not readable")
         return path
 
-    def a_file(self, *nodes: Union[Path, str], suffixes: Optional[typing.Set[str]] = None) -> Path:
-        """Gets a path of a test resource file under ``resources/``, ignoring suffix."""
-        path = Path(self.resource_dir, "resources", *nodes)
+    def a_file(self, *nodes: PathLike, suffixes: Optional[typing.Set[str]] = None) -> Path:
+        """
+        Gets a path of a test resource file under ``resources/``, ignoring suffix.
+
+        Args:
+            nodes: Path nodes under ``resources/``
+            suffixes: Set of acceptable suffixes; if None, all are accepted
+        """
+        path = Path(self._dir, "resources", *nodes)
         options = [
             p
             for p in path.parent.glob(path.stem + "*")
@@ -65,37 +100,41 @@ class MandosResources:
         except LookupError:
             raise MissingResourceError(f"Resource {path} missing") from None
 
-    def json(self, *nodes: Union[Path, str], suffix: Optional[str] = None) -> NestedDotDict:
+    def json(self, *nodes: PathLike, suffix: Optional[str] = None) -> NestedDotDict:
         """Reads a JSON file under ``resources/``."""
         path = self.path(*nodes, suffix=suffix)
-        data = orjson.loads(Path(path).read_text(encoding="utf8"))
+        data = orjson.loads(Path(path).read_text(encoding="utf8", errors="strict"))
         return NestedDotDict(data)
 
-    def json_dict(self, *nodes: Union[Path, str], suffix: Optional[str] = None) -> MutableMapping:
+    def json_dict(self, *nodes: PathLike, suffix: Optional[str] = None) -> MutableMapping:
         """Reads a JSON file under ``resources/``."""
         path = self.path(*nodes, suffix=suffix)
-        data = orjson.loads(Path(path).read_text(encoding="utf8"))
+        data = orjson.loads(Path(path).read_text(encoding="utf8", errors="strict"))
         return data
 
-    def check_expired(self, path: Path, max_sec: Union[timedelta, int], what: str) -> bool:
-        if isinstance(max_sec, timedelta):
-            max_sec = max_sec.total_seconds()
-        # getting the mod date because creation dates are iffy cross-platform
-        # (in fact the Linux kernel doesn't bother to expose them)
-        when = datetime.fromtimestamp(path.stat().st_mtime)
-        delta_sec = (datetime.now() - when).total_seconds()
-        if delta_sec > max_sec:
-            delta_str = UnitTools.delta_time_to_str(Chars.narrownbsp)
-            if delta_sec > 60 * 60 * 24 * 2:
-                self.logger.warning(
-                    f"{what} may be {delta_str} out of date. [downloaded: {when.strftime('%Y-%m-%d')}]"
-                )
-            else:
-                self.logger.warning(
-                    f"{what} may be {delta_str} out of date. [downloaded: {when.strftime('%Y-%m-%d %H:%M:%s')}]"
-                )
+    def check_expired(
+        self, path: PathLike, max_sec: Union[timedelta, float], *, what: Optional[str] = None
+    ) -> bool:
+        """
+        Warns and returns True if ``path`` mod date is more than ``max_sec`` in the past.
+
+        Args:
+            path: A specific path to check
+            max_sec: Max seconds, or a timedelta
+            what: Substitute the path with this string in logging
+        """
+        path = Path(path)
+        what = str(path) if what is None else what
+        limit = max_sec if isinstance(max_sec, timedelta) else timedelta(seconds=max_sec)
+        now = datetime.now().astimezone()
+        then = FilesysTools.get_info(path).mod_or_create_dt
+        delta = now - then
+        if delta > limit:
+            delta_str = UnitTools.delta_time_to_str(delta, space=Chars.narrownbsp)
+            then_str = UnitTools.approx_time_wrt(now, then)
+            self._logger.warning(f"{what} may be {delta_str} out of date. [{then_str}]")
             return True
         return False
 
 
-__all__ = ["MandosResources"]
+__all__ = ["Resources"]